Refactored `EnumeratedCaseUnion` into `CaseUnion`
The new functionality uses attributes so that the resulting type does not need interface implementations that need to either implement a property of a specified name or offer an explicit interface implementation that could lead to the type being boxed.  The interface does not seem to be extremely useful when compared with these issues.